requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.7.0
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-gitlab==2.0.1
pytz==2019.3
regex==2020.2.20
requests==2.23.0
six==1.14.0
toml==0.10.0
twilio==6.35.5
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
